pong_contest/requirements.txt
Django>=4.0,<5.0
psycopg2-binary
python-dotenv
Pillow
pyjwt==2.8.0
